clarified naming to lobby + other stuff?
