+++ src/routes.js
@@ -3,7 +3,7 @@ import { Route, Switch, Redirect } from "react-router";
 import PhoneHome from "./PhoneHome";
 import App from "./App";
 import PictionaryHome from "./Game-Pic";
-import GameHome from "./GameHome";
+import Lobby from "./Lobby";
 
 export default class Routes extends Component {
   render() {
@@ -22,7 +22,7 @@ export default class Routes extends Component {
     } else {
       return (
         <Switch>
-          <Route path="/lobby" component={GameHome} />
+          <Route path="/lobby" component={Lobby} />
           <Route path="/browse" component={App} />
           <Redirect from="/" to="browse" />
         </Switch>
